Implemented basic async io for the library

+++ arch/types.h
@@ -4,4 +4,4 @@
 
 typedef uint16_t gpio_pin_t;
 
-typedef int64_t timestamp_t;
+typedef int32_t timestamp_t;
